% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Documentation for CountryPopulation}
\usage{
CountryPopulation(country)
}
\arguments{
\item{a}{An official Country's offical name}
}
\value{
A graph of a country's population from 1950 to 2020
}
\description{
CountryPopulation is a function that takes the input of a country and returns
a graph of that country's population from 1950 to 2020.
}
\examples{
CountryPopulation("United States of America")
CountryPopulation("Russia")
CountryPopulation("China")
CountryPopulation("United Kingdom")
}
