% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/World_Cupdoc.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{Documentation for World Cup Data Set}
\format{
A data frame with 22 observations with 4 columns.
\describe{
   \item{WorldCup}{The country and year that the subsequent columns contain
                   data for}
   \item{Matches}{Number of matches in total for that World Cup year}
   \item{Totalattendance}{Total attendance for all matches}
   \item{Averageattendance}{Average attendance for each match}
}
}
\usage{
World_Cup
}
\description{
This data set contains World Cup data from 1930 to 2022
}
\keyword{datasets}
