% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorldPopulation_datasetdoc.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{Documentation for World Population Data Set}
\format{
A data frame with 235 observations with 72 columns.
\describe{
   \item{Country}{The country that the subsequent columns have data for}
   \item{1950}{The Population of the country in 1950 in thousands}
   \item{1951}{The Population of the country in 1951 in thousands}
   \item{1952}{The Population of the country in 1952 in thousands}
   \item{1953}{The Population of the country in 1953 in thousands, every
               subsequent column will follow this pattern}
}
}
\usage{
WorldPopulation
}
\description{
This data set contains the populations of every United Nation recognized
country from 1950 to 2020.
}
\keyword{datasets}
