% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{Documentation for CountryPopulation}
\usage{
CountryPopulation(country)
}
\arguments{
\item{a}{An official Country's offical name}
}
\value{
A graph of a country's population from 1950 to 2020
}
\description{
CountryPopulation is a function that takes the input of a country and returns
a graph of that country's population from 1950 to 2020.You have to use the
United Nations recognized name for that country. In the examples below you
will notice that the graph for "Russia" did not work, this is because the
offical name for Russia is "Russian Federation."
}
\examples{
CountryPopulation("United States of America")
CountryPopulation("Russia")
CountryPopulation("China")
CountryPopulation("United Kingdom")
}
